refactor(router): renamed Verify flow to SetupMandate
